idea build and UI Designer settings change
The project is buildable again with idea, maven does not work (i don't know how it works)

The UI Designer code in the client app is generated into the source code its self.
That is because if it's not generated in the source code the app can be compiled only in intellij idea with the UI Designer plugin.